src/components: move MobileBottomPanel
